Get the Sandbox stuff working properly and rename everything so it's sexy af
