+++ QuartzSandbox/Source/CMakeLists.txt
@@ -1,7 +1,6 @@
 set(currentDir ${CMAKE_CURRENT_LIST_DIR})
-
-set(clientSources
-	${currentDir}/Client.cpp
+set(sandboxSources
+	${currentDir}/Sandbox.cpp
 
 	PARENT_SCOPE
 )
