+++ QuartzSandbox/assets/CMakeLists.txt
@@ -1,12 +1,12 @@
 cmake_minimum_required(VERSION 3.0)
 
-project(quartz-sandbox-assets)
+project(QuartzSandboxAssets)
 
 set(assetsPath ${CMAKE_CURRENT_SOURCE_DIR})
 
-file(GLOB shaders ${assetsPath}/Shaders/*.vert ${assetsPath}/Shaders/*.frag)
-file(GLOB scripts ${assetsPath}/Scripts/*.lua)
-file(GLOB textures ${assetsPath}/Textures/*.png)
+file(GLOB shaders ${assetsPath}/shaders/*.vert ${assetsPath}/shaders/*.frag)
+file(GLOB scripts ${assetsPath}/scripts/*.lua)
+file(GLOB textures ${assetsPath}/textures/*.png)
 
 source_group("Shaders" FILES ${shaders})
 source_group("Scripts" FILES ${scripts})
